seperate game info

+++ MK11Hook/code/mkcamera.cpp
@@ -230,7 +230,7 @@ void MKCamera::HookedSetRotation(FRotator * rot)
 				else
 					GetObj(PLAYER1)->GetBoneRot("Head", &headRot);
 
-				rot->Pitch = headRot.Pitch + TheMenu->m_fAdjustCustomHeadCameraY;
+				rot->Pitch = headRot.Pitch + (int)TheMenu->m_fAdjustCustomHeadCameraY;
 				rot->Yaw = 16000 + headRot.Yaw;
 				rot->Roll = headRot.Roll / 100;
 				TheMenu->camRot = *rot;
